moved avr specifics to a seperate file, added the project settings on top of features.h
